+++ bionty/__init__.py
@@ -21,7 +21,7 @@ Entity collections:
 .. autosummary::
    :toctree: .
 
-   organism
+   species
 """
 
 from . import _version
@@ -31,7 +31,7 @@ __version__ = _version.get_versions()["version"]
 # bio entities
 from .gene import Gene
 from .protein import Protein
-from .taxon import Taxon, organism
+from .species import Taxon, species
 from .celltype import Celltype
 from .disease import Disease
 from .tissue import Tissue
