+++ bionty/species/__init__.py
@@ -1,2 +1,2 @@
-from ._core import Taxon, organism
+from ._core import Taxon, species
 from ._query import update_species_table, Mytaxon
